fix / SCROLLBAR_WIDTH returned value
when on an iOS device and the scrollbar pseudo element was present with
an explicit width value, (e.g: `::-webkit-scrollbar {width: 5px}`) the
SCROLLBAR_WIDTH variable will hold `5` instead of `0` (which is the value
we expect in order to prevent the custom-scrollbars), so, as a result the
custom-scrollbars were present, but they sould not.

to fix this, we are adding the `::-webkit-scrollbar` pseudo element on
the temporal element we use to measure the scrollbar-width setting its
`width` property to the `initial` value. This will return `0` on iOS
devices.