fix / hide custom scrollbar elements when prevented
there was an issue when the scrollbars markup was already declared on
the document, the scrollbars were customized with a background color and
the custom-scrollbars functionality was prevented. As a result we will
see the scrollbar elements (with its custom background color).

to fix this, we are hidding the custom scrollbars via CSS when our
scrollbar-width check returns 0. we also renamed the `gm-smooth-scroll`
class selector to `gm-prevented` to be more explicit of what it does
now.